requirements.txt
matplotlib==3.3.4
torchvision==0.8.2
torch==1.7.1+cu110
seaborn==0.11.1
rational_activations.egg==info
numpy==1.20.1
pandas==1.2.3
scikit_learn==0.24.1
tensorboard==2.4.1
utils==1.0.1

scikit-learn~=0.24.1
PyYAML~=3.12